refactor: remove story brief backward-compatibility facades
